+++ R/groupMSA.R
@@ -45,6 +45,8 @@ groupMSA <- function(groups, mpath = "intermediate_files/msa/",
 
   #loop through each group
   num <- max(groups$Group)
+
+
   if(num < 1){stop("Error: no input groups specified.")}
   num.length <- nchar(num)
 
@@ -131,7 +133,9 @@ groupMSA <- function(groups, mpath = "intermediate_files/msa/",
 
       }
     }
+
   }
+  close(pb)
 
   #merge pdfs
   if(pdflatex & pdftk){